// cctv-count-main/src/app/core/interfaces/station-data.interface.ts
export interface StationData {
    id?: string;
    station_name?: string;
    st_id: string;
    st_name: string;
    lat: string;
    lon: string;
  }

  export interface UpdateStationData {
    id?: string;
    station_name?: string;
    st_id?: string;
    st_name: string;
    lat: string;
    lon: string;
  }
  